close #18 refactor project
