+++ index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link href="/dist/styles.css" rel="stylesheet" />
-    <title>Vite + React</title>
+    <title>Schedulify</title>
   </head>
   <body>
     <div id="root"></div>
