+++ src/components/Card.jsx
@@ -1,9 +1,11 @@
 import React from "react";
-import {Link} from "react-router-dom";
-const Card = ({gen,specialize,group}) => {
+import { Link } from "react-router-dom";
+const Card = ({ gen, specialize, group }) => {
   return (
     <div className="w-[370px] h-[200px] shadow-md rounded-md p-4 relative">
-      <p className="text-lg font-semibold">Gen:{gen}</p>
+      <p className="text-lg font-semibold">
+        Generation: {gen < 10 ? `0${gen}` : gen}
+      </p>
 
       <div className="flex items-center mt-1">
         <span className="text-lg font-medium">{specialize}</span>
@@ -12,7 +14,7 @@ const Card = ({gen,specialize,group}) => {
       <p className="font-medium text-lg">
         Group: <span className="font-normal">{group}</span>
       </p>
-      <Link to="/myschedule" >
+      <Link to="/schedule">
         <div className="absolute bottom-2 right-2 text-white bg-[#284BAD] font-bold px-3 py-1 rounded-md hover:bg-[#1a306f]">
           View
         </div>
