+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,5 +1,6 @@
 import React, { useState } from "react";
 import { IoEyeOffOutline, IoEyeOutline } from "react-icons/io5";
+import { Link } from "react-router-dom";
 
 const LoginPage = () => {
   const [showPassword, setShowPassword] = useState(false);
@@ -8,8 +9,12 @@ const LoginPage = () => {
     <div className="flex flex-col justify-center items-center min-h-screen bg-[#F3F4F6] px-4 md:px-0">
       {/* Login Container */}
       <div className="bg-white rounded-2xl w-full max-w-sm md:max-w-md lg:max-w-lg p-6 md:p-8 shadow-lg">
-        <h1 className="text-2xl font-bold mb-2 text-center">Login to Schedulify</h1>
-        <p className="text-gray-500 mb-6 text-center">Login to stay connected</p>
+        <h1 className="text-2xl font-bold mb-2 text-center">
+          Login to Schedulify
+        </h1>
+        <p className="text-gray-500 mb-6 text-center">
+          Login to stay connected
+        </p>
 
         {/* Form */}
         <div className="flex flex-col gap-5">
@@ -43,7 +48,11 @@ const LoginPage = () => {
                 className="absolute right-3 top-3 text-gray-500 hover:text-gray-700"
                 onClick={() => setShowPassword(!showPassword)}
               >
-                {showPassword ? <IoEyeOffOutline size={20} /> : <IoEyeOutline size={20} />}
+                {showPassword ? (
+                  <IoEyeOffOutline size={20} />
+                ) : (
+                  <IoEyeOutline size={20} />
+                )}
               </button>
             </div>
           </div>
@@ -52,25 +61,26 @@ const LoginPage = () => {
           <div className="flex flex-col md:flex-row justify-between items-center text-sm">
             <div className="flex items-center gap-2">
               <input type="checkbox" id="rememberMe" className="w-4 h-4" />
-              <label htmlFor="rememberMe" className="cursor-pointer">Remember me</label>
+              <label htmlFor="rememberMe" className="cursor-pointer">
+                Remember me
+              </label>
             </div>
-            <p className="text-blue-500 hover:underline cursor-pointer">Forgot Password?</p>
+            <p className="text-blue-500 hover:underline cursor-pointer">
+              Forgot Password?
+            </p>
           </div>
 
           {/* Sign-In Button */}
           <div className="flex justify-center">
-            <button
-              type="submit"
-              className="bg-[#284BAD] text-white w-48 py-3 rounded-md hover:bg-blue-600 transition"
-            >
-              Sign In
-            </button>
-          </div>
-
-          {/* Register Link */}
-          <div className="flex justify-center text-sm">
-            <p>Don't have an account?</p>
-            <a href="#" className="text-[#284BAD] ml-1 hover:underline">Register</a>
+            <Link to="/">
+              <button
+                type="submit"
+                className="bg-[#284BAD] text-white w-48 py-3 rounded-md hover:bg-blue-600 transition"
+              >
+                Sign In
+              </button>
+            </Link>
+            {/* Sign-In Button */}
           </div>
         </div>
       </div>
