+++ src/pages/RequestPage/RequestPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const RequestPage = () => {
+  return (
+    <div>RequestPage</div>
+  )
+}
+
+export default RequestPage
