+++ src/pages/SchedulePage/SchedulePage.jsx
@@ -1,8 +1,9 @@
+
 import React, { useEffect, useState } from "react";
-import Table from "../../Components/ScheduleTable";
+import Table from "../../components/ScheduleTable";
 import { fetchSchedule } from "../../service/ScheduleData";
 
-const MySchedulePage = () => {
+const SchedulePage = () => {
   const [schedule, setSchedule] = useState([]);
 
   useEffect(() => {
@@ -27,7 +28,7 @@ const MySchedulePage = () => {
 
   return (
     <main className="max-w-screen-xl mx-auto p-4 flex flex-col gap-7 mt-5">
-      <h1 className="text-3xl font-bold">My Schedule</h1>
+      <h1 className="text-3xl font-bold">Schedule</h1>
       <Table headers={headers} rows={schedule} />
       <h2 className="text-2xl font-medium">Today's Classes:</h2>
       <ul className="list-disc pl-5">
@@ -42,4 +43,4 @@ const MySchedulePage = () => {
   );
 };
 
-export default MySchedulePage;
+export default SchedulePage;
