+++ src/service/HomepageData.js
@@ -2,13 +2,15 @@ export const fetchHomepage = async () => {
   return [
     { id: 1, Gen: 9, Specialize: "Software Engineer", Group: 2 },
     { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 3 },
-    { id: 3, Gen: 9, Specialize: "Product Manager", Group: 2 },
-    { id: 4, Gen: 11, Specialize: "UI/UX Designer", Group: 1 },
-    { id: 5, Gen: 10, Specialize: "Backend Developer", Group: 4 },
-    { id: 6, Gen: 12, Specialize: "Frontend Developer", Group: 5 },
-    { id: 7, Gen: 9, Specialize: "Business Analyst", Group: 3 },
-    { id: 8, Gen: 11, Specialize: "DevOps Engineer", Group: 4 },
-    { id: 9, Gen: 12, Specialize: "Full Stack Developer", Group: 1 },
-    { id: 10, Gen: 10, Specialize: "Machine Learning Engineer", Group: 2 },
+    { id: 1, Gen: 9, Specialize: "Software Engineer", Group: 4 },
+    { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 5 },
+    { id: 1, Gen: 9, Specialize: "Software Engineer", Group: 2 },
+    { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 3 },
+    { id: 1, Gen: 9, Specialize: "Software Engineer", Group: 2 },
+    { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 3 },
+    { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 3 },
+    { id: 1, Gen: 9, Specialize: "Software Engineer", Group: 2 },
+    { id: 2, Gen: 10, Specialize: "Data Scientist", Group: 3 },
+  
   ]
 };
